+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Changed
 - Client version updated on [5.1.15](https://github.com/reportportal/client-java/releases/tag/5.1.15), by @HardNorth
+- `NodeInfoUtils` class refactored to not use synchronized Map, by @HardNorth 
 
 ## [5.1.0]
 ### Added